- moved the daterange - made constants from strings
